+++ photon-core/src/main/java/org/photonvision/vision/processes/VisionRunner.java
@@ -38,6 +38,7 @@ public class VisionRunner {
     private final Supplier<CVPipeline> pipelineSupplier;
     private final Consumer<CVPipelineResult> pipelineResultConsumer;
     private final QuirkyCamera cameraQuirks;
+    private final VisionModule parentModule;
 
     private long loopCount;
 
@@ -53,15 +54,18 @@ public class VisionRunner {
             FrameProvider frameSupplier,
             Supplier<CVPipeline> pipelineSupplier,
             Consumer<CVPipelineResult> pipelineResultConsumer,
-            QuirkyCamera cameraQuirks) {
+            QuirkyCamera cameraQuirks,
+            VisionModule parentModule) {
         this.frameSupplier = frameSupplier;
         this.pipelineSupplier = pipelineSupplier;
         this.pipelineResultConsumer = pipelineResultConsumer;
         this.cameraQuirks = cameraQuirks;
+        this.parentModule = parentModule;
 
         visionProcessThread = new Thread(this::update);
         visionProcessThread.setName("VisionRunner - " + frameSupplier.getName());
         logger = new Logger(VisionRunner.class, frameSupplier.getName(), LogGroup.VisionModule);
+        parentModule.processSettingChanges();
     }
 
     public void startProcess() {
@@ -70,6 +74,7 @@ public class VisionRunner {
 
     private void update() {
         while (!Thread.interrupted()) {
+            parentModule.processSettingChanges();
             var pipeline = pipelineSupplier.get();
 
             // Tell our camera implementation here what kind of pre-processing we need it to be doing
